✨ feat: implement layout and metadata generation for user spaces
- Added a new layout component for user spaces that includes default
  frame and metadata generation.
- Enhanced metadata handling to include user-specific details and
  fallback options.
