+++ src/app/(spaces)/s/[handle]/page.tsx
@@ -4,10 +4,8 @@ import ProfileSpace, {
   UserDefinedSpacePageProps,
 } from "./ProfileSpace";
 import SpaceNotFound from "@/app/(spaces)/SpaceNotFound";
-import { Metadata } from "next/types";
-import { getUserMetadataStructure } from "@/common/lib/utils/userMetadata";
 import { unstable_noStore as noStore } from 'next/cache';
-import { WEBSITE_URL } from "@/constants/app";
+
 const loadUserSpaceData = async (
   handle: string,
   tabNameParam?: string,
@@ -44,63 +42,6 @@ const loadUserSpaceData = async (
 
   return { spaceOwnerFid, spaceOwnerUsername, spaceId, tabName };
 };
-
-export async function generateMetadata({
-  params: { handle, tabName: tabNameParam },
-}): Promise<Metadata> {
-  if (!handle) {
-    return {};
-  }
-
-  const userMetadata = await getUserMetadata(handle);
-  if (!userMetadata) {
-    return {};
-  }
-
-  // Process tabName parameter if it exists
-  const tabName = tabNameParam ? decodeURIComponent(tabNameParam) : undefined;
-  
-  // Create Frame metadata for Farcaster with the correct path
-  const frameUrl = tabName 
-    ? `${WEBSITE_URL}/s/${handle}/${encodeURIComponent(tabName)}`
-    : `${WEBSITE_URL}/s/${handle}`;
-    
-  const displayName = userMetadata?.displayName || userMetadata?.username || handle;
-  
-  const spaceFrame = {
-    version: "next",
-    imageUrl: `${WEBSITE_URL}/images/nounspace_og.png`,
-    button: {
-      title: `Visit ${displayName}'s Space`,
-      action: {
-        type: "launch_frame",
-        url: frameUrl,
-        name: `${displayName}'s Nounspace`,
-        splashImageUrl: `${WEBSITE_URL}/images/nounspace_logo.png`,
-        splashBackgroundColor: "#FFFFFF",
-      }
-    }
-  };
-
-  const baseMetadata = getUserMetadataStructure(userMetadata);
-  
-  // Type-safe way to add frame metadata
-  const metadataWithFrame = {
-    ...baseMetadata,
-    title: `${displayName}'s Space | Nounspace`,
-    description: userMetadata?.bio || 
-      `${displayName}'s customized space on Nounspace, the customizable web3 social app built on Farcaster.`,
-  };
-  
-  // Add the fc:frame metadata
-  if (!metadataWithFrame.other) {
-    metadataWithFrame.other = {};
-  }
-  
-  metadataWithFrame.other['fc:frame'] = JSON.stringify(spaceFrame);
-  
-  return metadataWithFrame;
-}
 
 const ProfileSpacePage = async ({
   params: { handle, tabName: tabNameParam },
